Moved DateAdded to WebMediaItem interface Added banner and fanart property to WebTVEpisodeBasic

+++ Branches/MPExtended-MEF/Services/MPExtended.Services.MediaAccessService.Interfaces/SortInterfaces.cs
@@ -27,6 +27,6 @@ namespace MPExtended.Services.MediaAccessService.Interfaces
 
     public interface IRatingSortable
     {
-        int Rating { get; set; }
+        float Rating { get; set; }
     }
 }
